requirements.txt
python-telegram-bot==20.8
openai>=1.12.0
python-dotenv==1.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiofiles==23.2.0
aiohttp==3.9.1
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.0
structlog==23.2.0
tenacity==8.2.3